MAJOR STRUCTURAL CHANGES IN BOT
Major major changes in the Bot section now allow for PROPER history... PROPER navigation.. all the code I've been able to come up with in the last several weeks in order to avoid the flaws of my spaghetti code (which honestly worked 90% pretty well) is now mostly useless... because I've found a way to make the navigation properly in the Bot section with components mounting and unmounting properly with a proper sequence and history.

MORE WORK needs to be done, a lot of functions will be useless, and a bit of CSS is needed to tidy up the Bot section.

But great progress today overall